Creates a mock cdm_reference from a preexisting cdm_reference object (it can be empty) and a list of tables. It makes sure that all the observations are within observation and all individuals are present in person table.

Creates a mock cdm_reference from a preexisting cdm_reference object (it can be empty) and a list of tables. It makes sure that all the observations are within observation and all individuals are present in person table.

Usage

mockCdmFromTables(cdm = mockCdmReference(), tables = list(), seed = NULL)

Arguments

cdm

A cdm_reference object

tables

List of named tables. They can be either omop standard tables or cohort tables.

seed

An integer to fix the seed of random numbers. If NULL seed is not set.

Value

a cdm object

Examples

# \donttest{
library(omock)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

cohort <- tibble(
  cohort_definition_id = c(1, 1, 2, 2, 1, 3, 3, 3, 1, 3),
  subject_id = c(1, 4, 2, 3, 5, 5, 4, 3, 3, 1),
  cohort_start_date = as.Date(c(
    "2020-04-01", "2021-06-01", "2022-05-22", "2010-01-01", "2019-08-01",
    "2019-04-07", "2021-01-01", "2008-02-02", "2009-09-09", "2021-01-01"
  )),
  cohort_end_date = cohort_start_date
)

cdm <- mockCdmFromTables(tables = list(cohort = cohort))

cdm
#> 
#> ── # OMOP CDM reference (local) of mock database ───────────────────────────────
#> • omop tables: person, observation_period, cdm_source, concept, vocabulary,
#> concept_relationship, concept_synonym, concept_ancestor, drug_strength
#> • cohort tables: cohort
#> • achilles tables: -
#> • other tables: -
cdm$cohort
#> # A tibble: 10 × 4
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <dbl>      <dbl> <date>            <date>         
#>  1                    1          1 2020-04-01        2020-04-01     
#>  2                    1          4 2021-06-01        2021-06-01     
#>  3                    2          2 2022-05-22        2022-05-22     
#>  4                    2          3 2010-01-01        2010-01-01     
#>  5                    1          5 2019-08-01        2019-08-01     
#>  6                    3          5 2019-04-07        2019-04-07     
#>  7                    3          4 2021-01-01        2021-01-01     
#>  8                    3          3 2008-02-02        2008-02-02     
#>  9                    1          3 2009-09-09        2009-09-09     
#> 10                    3          1 2021-01-01        2021-01-01     
cdm$person
#> # A tibble: 5 × 18
#>   person_id gender_concept_id year_of_birth month_of_birth day_of_birth
#> *     <dbl>             <dbl>         <dbl>          <dbl>        <int>
#> 1         1              8507          2011             11           18
#> 2         2              8532          2010             11            8
#> 3         3              8507          1989              7           24
#> 4         4              8507          2013              8           21
#> 5         5              8532          2012              5            9
#> # ℹ 13 more variables: birth_datetime <date>, race_concept_id <dbl>,
#> #   ethnicity_concept_id <dbl>, location_id <dbl>, person_source_value <chr>,
#> #   gender_source_value <chr>, gender_source_concept_id <int>,
#> #   race_source_value <chr>, race_source_concept_id <int>,
#> #   ethnicity_source_value <chr>, ethnicity_source_concept_id <int>,
#> #   provider_id <int>, care_site_id <int>

# }