Generates a mock death table and integrates it into an existing CDM object.

This function simulates death records for individuals within a specified cohort. It creates a realistic dataset by generating death records according to the specified number of records per person. The function ensures that each death record is associated with a valid person within the observation period to maintain the integrity of the data.

Usage

mockDeath(cdm, recordPerson = 1, seed = 1)

Arguments

cdm

A `cdm_reference` object that must already include 'person' and 'observation_period' tables. This object is the base CDM structure where the death data will be added. It is essential that the 'person' and 'observation_period' tables are populated as they provide necessary context for generating death records.

recordPerson

An integer specifying the expected number of death records to generate per person. This parameter helps simulate varying frequencies of death occurrences among individuals in the cohort, reflecting the variability seen in real-world medical data. Typically, this would be set to 1 or 0, assuming most datasets would only record a single death date per individual if at all.

seed

An optional integer used to set the seed for random number generation, ensuring reproducibility of the generated data. If provided, it allows the function to produce the same results each time it is run with the same inputs. If 'NULL', the seed is not set, which can result in different outputs on each run.

Value

Returns the modified `cdm` object with the new 'death' table added. This table includes the simulated death data for each person, ensuring that each record is linked correctly to individuals in the 'person' table and falls within valid observation periods.

Examples

# \donttest{
library(omock)

# Create a mock CDM reference and add death records
cdm <- mockCdmReference() |> mockPerson() |> mockObservationPeriod() |>
      mockDeath(recordPerson = 1)

# View the generated death data
print(cdm$death)
#> # A tibble: 10 × 3
#>    person_id death_date death_type_concept_id
#>  *     <int> <date>                     <dbl>
#>  1         9 2009-09-17                     1
#>  2         4 2013-12-31                     1
#>  3         7 2014-04-09                     1
#>  4         1 2013-06-29                     1
#>  5         2 2003-06-15                     1
#>  6         5 2013-02-10                     1
#>  7         3 2015-10-11                     1
#>  8        10 2019-12-26                     1
#>  9         6 2015-04-15                     1
#> 10         8 2009-03-10                     1
# }