Function to generate drug exposure table

Function to generate drug exposure table

Usage

mockDrugExposure(cdm, recordPerson = 1, seed = 1)

Arguments

cdm

the CDM reference into which the synthetic cohort will be added

recordPerson

The expected number of records per person within each cohort. This can help simulate the frequency of observations for individuals in the cohort.

seed

A random seed to ensure reproducibility of the generated data.

Value

A cdm reference with the mock tables

Examples

library(omock)

cdm <- mockCdmReference() |> mockPerson() |> mockObservationPeriod() |>
mockDrugExposure()

cdm$drug_exposure
#> # A tibble: 60 × 6
#>    drug_concept_id person_id drug_exposure_start_date drug_exposure_end_date
#>  *           <dbl>     <int> <date>                   <date>                
#>  1              10         9 1990-11-22               1994-03-05            
#>  2              10         4 1993-08-23               2000-05-06            
#>  3              10         7 2014-03-12               2014-03-15            
#>  4              10         1 2005-05-28               2007-06-05            
#>  5              10         2 1999-09-09               2001-05-19            
#>  6              10         7 2014-03-13               2014-03-20            
#>  7              10         2 2000-03-22               2000-10-18            
#>  8              10         3 2015-02-02               2015-06-03            
#>  9              10         1 2002-09-17               2005-09-14            
#> 10              10         5 2012-04-04               2012-08-28            
#> # ℹ 50 more rows
#> # ℹ 2 more variables: drug_exposure_id <int>, drug_type_concept_id <dbl>