Generates a mock observation period table and integrates it into an existing CDM object.

This function simulates observation periods for individuals based on their date of birth recorded in the 'person' table of the CDM object. It assigns random start and end dates for each observation period within a realistic timeframe up to a specified or default maximum date.

Usage

mockObservationPeriod(cdm, seed = NULL)

Arguments

cdm

A `cdm_reference` object that must include a 'person' table with valid dates of birth. This object serves as the base CDM structure where the observation period data will be added. The function checks to ensure that the 'person' table is populated and uses the date of birth to generate observation periods.

seed

An optional integer used to set the seed for random number generation, ensuring reproducibility of the generated data. If provided, this seed allows the function to produce consistent results each time it is run with the same inputs. If 'NULL', the seed is not set, which can lead to different outputs on each run.

Value

Returns the modified `cdm` object with the new 'observation_period' table added. This table includes the simulated observation periods for each person, ensuring that each record spans a realistic timeframe based on the person's date of birth.

Examples

library(omock)

# Create a mock CDM reference and add observation periods
cdm <- mockCdmReference() |>
  mockPerson(nPerson = 100) |>
  mockObservationPeriod()

# View the generated observation period data
print(cdm$observation_period)
#> # A tibble: 100 × 5
#>    observation_period_id person_id observation_period_s…¹ observation_period_e…²
#>  *                 <int>     <int> <date>                 <date>                
#>  1                     1         1 1975-03-20             2015-05-31            
#>  2                     2         2 1995-01-25             2009-06-19            
#>  3                     3         3 1993-07-23             1999-09-06            
#>  4                     4         4 1982-06-09             2012-04-04            
#>  5                     5         5 1992-04-15             1992-05-16            
#>  6                     6         6 1992-05-09             2011-12-12            
#>  7                     7         7 2005-03-25             2009-05-13            
#>  8                     8         8 2015-06-03             2018-02-23            
#>  9                     9         9 2017-12-02             2019-11-09            
#> 10                    10        10 1972-09-08             1993-05-11            
#> # ℹ 90 more rows
#> # ℹ abbreviated names: ¹​observation_period_start_date,
#> #   ²​observation_period_end_date
#> # ℹ 1 more variable: period_type_concept_id <int>