Generates a mock procedure occurrence table and integrates it into an existing CDM object.

This function simulates condition occurrences for individuals within a specified cohort. It helps create a realistic dataset by generating condition records for each person, based on the number of records specified per person.The generated data are aligned with the existing observation periods to ensure that all conditions are recorded within valid observation windows.

Usage

mockProcedureOccurrence(cdm, recordPerson = 1, seed = NULL)

Arguments

cdm

A `cdm_reference` object that should already include 'person', 'observation_period', and 'concept' tables.This object is the base CDM structure where the procedure occurrence data will be added. It is essential that these tables are not empty as they provide the necessary context for generating condition data.

recordPerson

An integer specifying the expected number of condition records to generate per person.This parameter allows the simulation of varying frequencies of condition occurrences among individuals in the cohort, reflecting the variability seen in real-world medical data.

seed

An optional integer used to set the seed for random number generation, ensuring reproducibility of the generated data.If provided, it allows the function to produce the same results each time it is run with the same inputs.If 'NULL', the seed is not set, resulting in different outputs on each run.

Value

Returns the modified `cdm` object with the new 'condition_occurrence' table added. This table includes the simulated condition data for each person, ensuring that each record is within the valid observation periods and linked to the correct individuals in the 'person' table.

Examples

# \donttest{
library(omock)

# Create a mock CDM reference and add condition occurrences
cdm <- mockCdmReference() |>
  mockPerson() |>
  mockObservationPeriod() |>
  mockProcedureOccurrence(recordPerson = 2)

# View the generated condition occurrence data
print(cdm$procedure_occurrence)
#> # A tibble: 40 × 15
#>    procedure_concept_id person_id procedure_date procedure_end_date
#>  *                <int>     <int> <date>         <date>            
#>  1              4336464         7 1986-11-17     1987-05-10        
#>  2              4336464         8 1990-06-29     1992-06-17        
#>  3              4336464        10 2005-03-07     2005-09-30        
#>  4              4336464         9 2003-07-12     2005-03-31        
#>  5              4336464         8 1990-08-23     1992-10-25        
#>  6              4336464         7 1986-11-01     1987-04-22        
#>  7              4336464         4 1996-08-29     1997-06-15        
#>  8              4336464         9 2002-03-28     2004-04-20        
#>  9              4336464         2 1989-01-09     1990-03-21        
#> 10              4336464         4 1996-11-24     1997-04-02        
#> # ℹ 30 more rows
#> # ℹ 11 more variables: procedure_occurrence_id <int>,
#> #   procedure_type_concept_id <int>, procedure_datetime <dttm>,
#> #   modifier_concept_id <int>, quantity <int>, provider_id <int>,
#> #   visit_occurrence_id <int>, visit_detail_id <int>,
#> #   procedure_source_value <chr>, procedure_source_concept_id <int>,
#> #   modifier_source_value <chr>
# }